fix: backend quality improvements — deduplicate helpers, fix security & bugs
1. Extract shared Celery helpers into tasks/helpers.py (deduplicate _get_redis,
   _get_sync_db, _publish_progress, _update_job_db from 4 task files)
2. Remove duplicate API endpoints (bulk-tag, bulk-delete, collection frames)
3. Fix path traversal in frame/thumbnail serving with shared validate_file_path
4. Fix traceback parameter shadowing traceback module in on_task_failure
5. Update gap_detector to use GoesFrame model instead of legacy Image
   (band not channel, includes sector filtering)
6. Use proper sessionmaker with pooled engine for sync DB sessions
7. Add CompositeCreateRequest Pydantic model (replaces raw dict)
8. Change Animation.false_color from Integer to Boolean (idempotent migration)
9. Add pagination to frames export endpoint (limit/offset params)
10. Add pagination to collection frames endpoint (limit/offset params)
